// src/types/Planet.ts
interface Planet {
  id: string;
  name: string;
  diameter: number;
  surfaceArea: number;
  mass: string;
  distanceFromSun: number;
  naturalSatellites: number;
  rotationPeriod: string;
  translationPeriod: string;
  averageTemperature: number;
  description: string;
  color: string;
  textureUrl: string;
}

export default Planet; 